Extract SharedPaymentTokenConfirmationInterceptor.kt
